refactor: tag 도메인 분리
