+++ src/test/java/com/meoguri/linkocean/domain/bookmark/entity/BookmarkTest.java
@@ -18,6 +18,8 @@ import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
 import com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType;
 import com.meoguri.linkocean.domain.linkmetadata.entity.LinkMetadata;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
+import com.meoguri.linkocean.domain.tag.entity.Tag;
+import com.meoguri.linkocean.domain.tag.entity.Tags;
 
 class BookmarkTest {
