+++ src/test/java/com/meoguri/linkocean/domain/tag/entity/TagTest.java
@@ -1,6 +1,6 @@
-package com.meoguri.linkocean.domain.bookmark.entity;
+package com.meoguri.linkocean.domain.tag.entity;
 
-import static com.meoguri.linkocean.domain.bookmark.entity.Tag.*;
+import static com.meoguri.linkocean.domain.tag.entity.Tag.*;
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
